Entity Details View Level 1

+++ src/main/java/eu/fvsware/jpaexplorer/service/EmployeeRepository.java
@@ -3,5 +3,5 @@ package eu.fvsware.jpaexplorer.service;
 import eu.fvsware.jpaexplorer.model.Employee;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmployeeRepository extends CrudRepository<Employee, Long> {
+public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
 }
